+++ cmd/gen/renovate/command.go
@@ -11,8 +11,8 @@ import (
 
 const (
 	name        = "renovate"
-	description = "Generates Renovate config for go and docker dependencies (renovate.yml)."
-	example     = `  devctl gen renovate 
+	description = "Generates Renovate config for go and docker dependencies (renovate.json5)."
+	example     = `  devctl gen renovate
   devctl gen renovate --interval "after 9am on thursday" --reviewers giantswarm/team-firecracker`
 )