+++ pkg/gen/gen.go
@@ -84,7 +84,7 @@ func isRegenerable(path string) bool {
 		return true
 	case base == ".gitignore":
 		return true
-	case base == "renovate.json":
+	case base == "renovate.json" || base == "renovate.json5":
 		return true
 	case base == "dependabot.yml":
 		return true