+++ pkg/gen/input/renovate/internal/file/renovate.go
@@ -8,12 +8,12 @@ import (
 	"github.com/giantswarm/devctl/pkg/gen/input/renovate/internal/params"
 )
 
-//go:embed renovate.json.template
+//go:embed renovate.json5.template
 var createRenovateTemplate string
 
 func NewCreateRenovateInput(p params.Params) input.Input {
 	i := input.Input{
-		Path:         filepath.Join(p.Dir, "renovate.json"),
+		Path:         filepath.Join(p.Dir, "renovate.json5"),
 		TemplateBody: createRenovateTemplate,
 		TemplateData: map[string]interface{}{
 			"Interval": params.Interval(p),